OPL3 fork: hardware FM synthesis (opl3_fpga) + FM piano demo (v0.15.0)
Integrates Greg Taylor's opl3_fpga (LGPL, vendored in pocket_core/opl3/)
into the core on this fork branch:

- Runs entirely in the 12.288 MHz domain (pkg CLK_FREQ patched, divider
  248 -> 49548 Hz sample rate, 0.34% below the authentic 49716 Hz)
- Classic two-port register interface driven from the SoC: main_opl_cmd
  CSR {A[1:0],D[7:0]}, one bus write per CSR write (auto toggle)
- FM output (24-bit) mixed with the PCM stream in core_top, saturated,
  into sound_i2s. The CPU never synthesizes FM.
- HAL: opl_write(reg9, val) (2us paced), HAL_FEAT_FM in sys_caps
- sdk/fmdemo: an FM piano — d-pad+BAYX = one octave, L1/R1 octave
  shift, 2-op e-piano patch, one OPL3 channel per key. No PCM code.

SYNTHESIS BATTLE (documented for posterity): Quartus's optimizer
repeatedly swept the entire OPL3 claiming its sample outputs were
stuck at GND — provably falsely (it persisted even with the register
bus driven from raw controller pins). Countermeasures that did NOT
work: ic_n rewiring, clean db, PRESERVE_REGISTER via qsf or instance
altera_attribute. What worked: (* preserve *) on the opl_l/opl_r mix
registers in core_top — the first stone of the removal domino — which
forces the whole FM fan-in to be implemented. +767 registers, ALMs 30%,
timing clean. Hardware test pending: fmdemo must actually sing.

+++ soc/hal/hal.h
@@ -123,10 +123,13 @@ void      input_state(int player, hal_pad_t *out);                   // [BUILT]
 // Opinionated but general: Tyrian leans on opl_write; a Quake port ignores it and
 // uses the stream. The HAL offers the surface; each app uses the part it needs.
 
-// AdLib/OPL2 register write. The design's thesis: the CPU never synthesizes FM —
-// it just forwards the register writes its music engine already emits.
-// backed by: hardware OPL2 block MMIO port.                          [PLANNED]
-void      opl_write(uint8_t reg, uint8_t val);
+// OPL3 register write (reg 0x000-0x0FF bank 0, 0x100-0x1FF bank 1). The
+// design's thesis: the CPU never synthesizes FM — it just forwards the
+// register writes its music engine already emits. FM output is hardware-mixed
+// with the PCM stream. Remember: OPL3 mode needs reg 0x105 bit0, and each
+// channel's 0xC0 needs the L/R output bits (0x30).
+// backed by: opl3_fpga core + main_opl_cmd CSR.    [BUILT on the opl3 fork]
+void      opl_write(uint16_t reg, uint8_t val);
 
 // Fire-and-forget PCM sound effect on a voice (mono, any rate <= 48k; 4 voices;
 // ch = -1 picks a free one, returns the channel or -1). Games using voices call
